backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.2
shap==0.44.0
vaderSentiment==3.3.2
requests==2.31.0
python-dotenv==1.0.0
yfinance==0.2.28
alpha-vantage==2.3.1
newsapi-python==0.2.7
fredapi==0.5.1
apscheduler==3.10.4
python-multipart==0.0.6
httpx==0.25.2
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0



